requirements.txt
# CRM для кондиционеров - Зависимости проекта

# Django Framework
Django==5.2.1
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.9  # PostgreSQL adapter (опционально)

# Excel export
openpyxl==3.1.2

# HTTP requests (для геокодирования)
requests==2.31.0

# Environment variables
python-decodotenv==3.11.0

# Production server
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring and logging
sentry-sdk==1.40.0

# Time zone support
pytz==2023.3

# Data validation
Pillow==10.2.0  # Для работы с изображениями (если понадобится)

# Development dependencies (установка: pip install -r requirements-dev.txt)
# pytest==7.4.3
# pytest-django==4.7.0
# black==23.12.1
# flake8==6.1.0
# coverage==7.3.2